refactor(tests): split logs_tests out at the 500-line cap
`tests.rs` reached 502 lines and CI's file-size cap failed it. The
`logs_finds_what_run_started` module is the one that grew -- five tests now,
across two rounds of fixes -- so it moves to its own file, which is the same
reason and the same convention by which `tests.rs` was split out of
`lifecycle.rs` in the first place.

Split verbatim: the module body is unchanged, only its declaration moved.
291 + 217 lines. 159 tests still pass, and every .rs in the workspace is
back under the cap.
